refactor: organiza tipos de filtros e serviços para Coeq e Coaps, criando modelos compartilhados
