+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/index.tsx
@@ -1,3 +1,4 @@
 export { CoapsDataTable } from "./modules/CoapsDataTable";
+export type { AppliedFiltersCoaps } from "./modules/CoapsDataTable";
 export { CoeqDataTable } from "./modules/CoeqDataTable";
-export type { AppliedFiltersCoaps, AppliedFiltersCoeq } from "./model";
+export type { AppliedFiltersCoeq } from "./modules/CoeqDataTable";
