+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/modules/CoapsDataTable/index.tsx
@@ -26,9 +26,10 @@ import * as service from "./service";
 import { FiltersContext } from "@/features/acf/frontend/common/WithFilters/context";
 import type { AppliedFilters } from "@/features/acf/frontend/common/WithFilters/model";
 import { EmptyTableMessage } from "../common/EmptyTableMessage";
-import type { AppliedFiltersCoaps } from "../../model";
 
-// export type { AppliedFiltersCoaps } from "./model";
+import type { AppliedFiltersCoaps } from "./model";
+
+export type { AppliedFiltersCoaps } from "./model";
 
 const fetchPage = (
     session: Session | null,
