+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/modules/CoapsDataTable/model.ts
@@ -1,0 +1,6 @@
+//TODO: rever este import
+import type { SharedAppliedFilters } from "../common/SharedAppliedFilters";
+
+export type AppliedFiltersCoaps = SharedAppliedFilters & {
+    careTeamName: Array<string>;
+};
