+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/modules/CoeqDataTable/model.ts
@@ -1,0 +1,3 @@
+import type { SharedAppliedFilters } from "../common/SharedAppliedFilters";
+
+export type AppliedFiltersCoeq = SharedAppliedFilters;
