+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/modules/common/SharedAppliedFilters.ts
@@ -5,13 +5,9 @@ import type {
     CommunityHealthWorker,
 } from "@/features/acf/shared/diabetes/model";
 
-export type AppliedFiltersCoeq = {
+export type SharedAppliedFilters = {
     communityHealthWorker: Array<CommunityHealthWorker>;
     patientStatus: Array<PatientStatus>;
     conditionIdentifiedBy: ConditionIdentifiedBy | "";
     patientAgeRange: Array<PatientAgeRange>;
-};
-
-export type AppliedFiltersCoaps = AppliedFiltersCoeq & {
-    careTeamName: Array<string>;
 };
